Version 2.6

+++ src/main/java/com/proyecto/ecommerceRigobello/dto/ClientDto.java
@@ -8,7 +8,7 @@ public class ClientDto {
     private String name;
 
     private String lastname;
-    private Long years;
+    private Integer years;
 
     public void saveName (String name){
         this.name = name;
@@ -16,7 +16,7 @@ public class ClientDto {
     public void saveLastName (String lastname){
         this.lastname = lastname;
     }
-    public void saveYears(Long age) {
+    public void saveYears(Integer age) {
         this.years = age;
     }
